# EES-3437 Performance improvements to fetching subject metadata