--- a/src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Controllers/Api/Bau/BauCacheControllerTests.cs
+++ b/src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Controllers/Api/Bau/BauCacheControllerTests.cs
@@ -20,17 +20,88 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
     public class BauCacheControllerTests
     {
         [Fact]
-        public async Task ClearPrivateCache()
+        public async Task ClearPrivateCache_SingleValidPath()
         {
             var privateBlobStorageService = new Mock<IBlobStorageService>(Strict);
 
+            DeleteBlobsOptions options = null!;
+            var match = new CaptureMatch<DeleteBlobsOptions>(param => options = param);
+
             privateBlobStorageService
-                .Setup(s => s.DeleteBlobs(PrivateContent, default, default))
+                .Setup(
+                    s =>
+                        s.DeleteBlobs(PrivateContent, null, Capture.With(match)))
                 .Returns(Task.CompletedTask);
 
             var controller = BuildController(privateBlobStorageService: privateBlobStorageService.Object);
 
-            var result = await controller.ClearPrivateCache();
+            var result = await controller.ClearPrivateCache(
+                new ClearPrivateCachePathsViewModel
+                {
+                    Paths = SetOf("subject-meta")
+                }
+            );
+
+            result.AssertNoContent();
+
+            MockUtils.VerifyAllMocks(privateBlobStorageService);
+
+            var regex = Assert.IsType<Regex>(options.IncludeRegex);
+            Assert.Matches(regex, "releases/release-1/subject-meta/something");
+            Assert.DoesNotMatch(regex, "something/releases/release-1/subject-meta/something");
+            Assert.DoesNotMatch(regex, "releases/release-1/data-blocks/something");
+            Assert.DoesNotMatch(regex, "releases/release-1/invalid/something");
+            Assert.DoesNotMatch(regex, "releases/release-1/something");
+        }
+
+        [Fact]
+        public async Task ClearPrivateCacheReleases_AllValidPaths()
+        {
+            var privateBlobStorageService = new Mock<IBlobStorageService>(Strict);
+
+            DeleteBlobsOptions options = null!;
+            var match = new CaptureMatch<DeleteBlobsOptions>(param => options = param);
+
+            privateBlobStorageService
+                .Setup(
+                    s =>
+                        s.DeleteBlobs(PrivateContent, null, Capture.With(match)))
+                .Returns(Task.CompletedTask);
+
+            var controller = BuildController(privateBlobStorageService: privateBlobStorageService.Object);
+
+            var result = await controller.ClearPrivateCache(
+                new ClearPrivateCachePathsViewModel
+                {
+                    Paths = new HashSet<string>
+                    {
+                        "data-blocks",
+                        "subject-meta"
+                    }
+                }
+            );
+
+            result.AssertNoContent();
+
+            MockUtils.VerifyAllMocks(privateBlobStorageService);
+
+            var regex = Assert.IsType<Regex>(options.IncludeRegex);
+            Assert.Matches(regex, "releases/release-1/data-blocks/something");
+            Assert.Matches(regex, "releases/release-1/subject-meta/something");
+            Assert.DoesNotMatch(regex, "something/releases/release-1/data-blocks/something");
+            Assert.DoesNotMatch(regex, "something/releases/release-1/subject-meta/something");
+            Assert.DoesNotMatch(regex, "releases/release-1/invalid/something");
+            Assert.DoesNotMatch(regex, "releases/release-1/something");
+        }
+
+        [Fact]
+        public async Task ClearPrivateCache_Empty()
+        {
+            var privateBlobStorageService = new Mock<IBlobStorageService>(Strict);
+
+            var controller = BuildController(privateBlobStorageService: privateBlobStorageService.Object);
+
+            var result = await controller.ClearPrivateCache(new ClearPrivateCachePathsViewModel());
 
             result.AssertNoContent();
 
@@ -54,7 +125,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
             var controller = BuildController(publicBlobStorageService: publicBlobStorageService.Object);
 
             var result = await controller.ClearPublicCacheReleases(
-                new ClearCacheReleasePathsViewModel
+                new ClearPublicCacheReleasePathsViewModel
                 {
                     Paths = SetOf("subject-meta")
                 }
@@ -66,6 +137,8 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
 
             var regex = Assert.IsType<Regex>(options.IncludeRegex);
             Assert.Matches(regex, "publications/publication-1/releases/release-1/subject-meta/something");
+            Assert.DoesNotMatch(regex,
+                "something/publications/publication-1/releases/release-1/subject-meta/something");
             Assert.DoesNotMatch(regex, "publications/publication-1/releases/release-1/data-blocks/something");
             Assert.DoesNotMatch(regex, "publications/publication-1/releases/release-1/fast-track-results/something");
             Assert.DoesNotMatch(regex, "publications/publication-1/releases/release-1/invalid/something");
@@ -89,7 +162,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
             var controller = BuildController(publicBlobStorageService: publicBlobStorageService.Object);
 
             var result = await controller.ClearPublicCacheReleases(
-                new ClearCacheReleasePathsViewModel
+                new ClearPublicCacheReleasePathsViewModel
                 {
                     Paths = new HashSet<string>
                     {
@@ -108,6 +181,11 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
             Assert.Matches(regex, "publications/publication-1/releases/release-1/data-blocks/something");
             Assert.Matches(regex, "publications/publication-1/releases/release-1/subject-meta/something");
             Assert.Matches(regex, "publications/publication-1/releases/release-1/fast-track-results/something");
+            Assert.DoesNotMatch(regex, "something/publications/publication-1/releases/release-1/data-blocks/something");
+            Assert.DoesNotMatch(regex,
+                "something/publications/publication-1/releases/release-1/subject-meta/something");
+            Assert.DoesNotMatch(regex,
+                "something/publications/publication-1/releases/release-1/fast-track-results/something");
             Assert.DoesNotMatch(regex, "publications/publication-1/releases/release-1/invalid/something");
             Assert.DoesNotMatch(regex, "publications/publication-1/releases/release-1/something");
         }
@@ -119,7 +197,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
 
             var controller = BuildController(publicBlobStorageService: publicBlobStorageService.Object);
 
-            var result = await controller.ClearPublicCacheReleases(new ClearCacheReleasePathsViewModel());
+            var result = await controller.ClearPublicCacheReleases(new ClearPublicCacheReleasePathsViewModel());
 
             result.AssertNoContent();
 
@@ -140,7 +218,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
             var controller = BuildController(publicBlobStorageService: publicBlobStorageService.Object);
 
             var result = await controller.ClearPublicCacheTrees(
-                new ClearCacheTreePathsViewModel
+                new ClearPublicCacheTreePathsViewModel
                 {
                     Paths = SetOf("publication-tree-any-data.json"),
                 }
@@ -168,7 +246,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
             var controller = BuildController(publicBlobStorageService: publicBlobStorageService.Object);
 
             var result = await controller.ClearPublicCacheTrees(
-                new ClearCacheTreePathsViewModel
+                new ClearPublicCacheTreePathsViewModel
                 {
                     Paths = new HashSet<string>
                     {
@@ -196,7 +274,7 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin.Tests.Controllers.Api
 
             var controller = BuildController(publicBlobStorageService: publicBlobStorageService.Object);
 
-            var result = await controller.ClearPublicCacheTrees(new ClearCacheTreePathsViewModel());
+            var result = await controller.ClearPublicCacheTrees(new ClearPublicCacheTreePathsViewModel());
 
             result.AssertNoContent();