The motivation behind this PR is that underlying datasets containing school level data are soon to be released in EES and 
the existing location wizard step in the table builder is going to be slow to use since it currently presents the full list of location options available. There could be 20k+ options.

We could implement a dedicated back-end location search and remove the dependency on all the locations being supplied upfront in the subject metadata response, but some rework would be required to restore data blocks and fast tracks in the table builder. See the notes of EES-3433 which has investigated this for more details.

Before we do this, we'd like to know if the existing client side location search would be sufficient if we can make the current subject meta data response more performant, and alter the front-end to not initially render any options until a search is performed.

Here's a current view of the client side search option:
![image](https://user-images.githubusercontent.com/4147126/173878663-bfbc5afe-dd08-4535-8653-253d88147ac4.png)

A separate PR will be raised for the front-end improvements, see EES-3452.

This PR has investigated ways of making the subject meta data response more performant and come up with four changes:

- Enable response compression
- Optimise the `EnumToEnumValueConverter` which provides a way of converting from `GeographicLevel` enum value strings in  to enum model values for each database location row returned.
- Optimise the way we deduplicate location view models.
- Add caching of unfiltered subject meta data requests to the Admin.

### Before and after

These changes were tested in my local development environment running in non-debug mode with an underlying dataset containing 22k schools and times averaged over a sample of requests.

#### Response size
Before: 2.3Mb
After: 776Kb

#### Response times
- Before with no changes: 2239ms
- Changing only `EnumToEnumValueConverter` only: 1425ms
- Changing only `DeduplicateLocationViewModels`:  1146ms
- Changing `EnumToEnumValueConverter` and `DeduplicateLocationViewModels`: 448ms
- Changing `EnumToEnumValueConverter` and `DeduplicateLocationViewModels` and enabling compression: 390ms
- Cached subject meta data and compression: 203ms

Enabling compression on it's own had very little or no effect in the local environment.

It's possible that the change to optimise the `EnumToEnumValueConverter` will also have had an positive impact on table result response times since the `TimeIdentifier` field of `Observation` also uses the same conversion.

### Other changes

- `ReleaseServiceTests` : Tidies up some of the mock verifications and fixes a few issues with missing or unnecessary setup of mock behaviour seen after enabling `strict` mode on all of the mocked services.
- Changes the layout of the private blob cache to no longer cache release content in a hierarchy by publication. The top level directory in the private cache changes now to be `releases` rather than `publications`. To cache subject meta data we were either going to need a lookup of the publication id or add it into the request, just for the purpose of creating a cache path containing the publication, which seems unnecessary.
- Adds more granularity to `BauCacheController.ClearPrivateCache` to provide the option of clearing data-blocks or subject meta data, or both. Previously we only cached data blocks so there was no option.
- Deletes cached data blocks and subject meta data from the private  storage container for the previous release version when publishing release amendments.