@benoutram 2022-06-16
> The only thing I'm unsure about is whether the new cache files are being invalidated everywhere they need to be - I don't know enough about this part of the code to be sure about it. But even if that was the case, it would only happen on admin and would be fixed quickly, so I don't think it's a huge concern.

I've done a fair bit of testing around this. I tested deleting files, replacing files, cancelling amendments, changing the filters and indicators ordering, and publishing an amendment of the release. Invalidating the private cache of the previous version when amending a releases was added by this PR so it now invalidates cached data blocks as well as subject metadata.