@benoutram 2022-06-16
> Nice to see performance improvements like this! Just wondering if we need to update the confluence page related to clearing public/private cache trees?

Thanks @luke-h1 I've made a change to the Confluence page now for this change. I didn't even know that page existed!

I also spotted that it was out of date in terms of clearing the publication trees, based on this error that I saw when testing it out:

```
"Field must contain only values from: publication-tree.json, methodology-tree.json"
```

That also led to me adding an extra commit to this PR to remove a reference to some of those old paths still mentioned in the unit tests.